backend/requirements.txt
Flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
cryptography==41.0.4
pyjwt==2.8.0
python-dotenv==1.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1
gunicorn==21.2.0
eventlet==0.39.1